DEFCT17:
Link tags in the form of {@link #someLocalAttrib} that appear in class
docBlocks dont work. 

This is caused because the description of the docBlock is processed in
processClass() before context.addDocClass(cls) is called which means when
processLink() is called with a link similar to above the current class is
Null or Invlaid when processLink() tries to call
context.getCurrentClass().className.

Also MEMBER_REFERENCE_TPL is missing a class reference.

Examples of this can be found in the class docBlocks of:
Ext.Updater
Ext.Component

Attached is a small patch to move processing of descriptions after the
current class is registered with the context. It also fixes the
MEMBER_REFERENCE_TPL.


FIXED by jsuhre